% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{matcher}
\alias{matcher}
\title{Build a matcher}
\usage{
matcher(path, opts)
}
\arguments{
\item{path}{A layer path.}

\item{opts}{Arguments passed to \code{pater::match()}.}
}
\value{
A function to match against request paths.
}
\description{
Build a matcher
}
\keyword{internal}
